Raise an exception when timestamp is malformed
Previously, when a provided timestamp didn't match the regex format, it
would be configured as `nil`. This was nonsensical since a `nil`
timestamp is unusable, which allowed for subtle bugs to creep in.

Instead, we should alert the user as soon as an improper timestamp is
encountered by raising an informative exception.

The documentation has also been updated to reflect the expectations for
tiemstamp format.

Fixes #140.